+++ docs/user_guide/changelog.rst
@@ -28,6 +28,7 @@ New features
 - Support for ``torch.nn.utils.weight_norm``
 - Support for ``torch.randperm``
 - Support for ``torch.nn.functional.cosine_similarity`` and ``torch.nn.CosineSimilarity``
+- Support for ``torch.all``, ``torch.any``, ``torch.Tensor.all`` and ``torch.Tensor.any``
 
 API changes
 -----------